Subscription Id - long to int
Subscription id changed from long to int as per new api.

CRs-Fixed: 821673

+++ telecomm/java/android/telecom/TelecomManager.java
@@ -929,7 +929,7 @@ public class TelecomManager {
      * Returns current active subscription.
      * @hide
      */
-    public long getActiveSubscription() {
+    public int getActiveSubscription() {
         try {
             if (isServiceConnected()) {
                 return getTelecomService().getActiveSubscription();
@@ -944,7 +944,7 @@ public class TelecomManager {
      * switches to other active subscription.
      * @hide
      */
-    public void switchToOtherActiveSub(long subId) {
+    public void switchToOtherActiveSub(int subId) {
         try {
             if (isServiceConnected()) {
                 getTelecomService().switchToOtherActiveSub(subId);
